+++ CHANGES.rst
@@ -10,7 +10,11 @@ https://zope.readthedocs.io/en/2.13/CHANGES.html
 4.5.1 (unreleased)
 ------------------
 
+
 - Make "chameleon-zope context wrapping" more faithful
+
+- Let "unicode conflict resolution" work for all templates (not just
+  ``ZopePageTemplate``).
 
 - Update dependencies to the latest releases that still support Python 2.